Refresh TravisCI test runner
Configure the distribution to be Ubuntu 14.04 Trusty which is tagged beta.
Manually install Riak 2.1.4 since it is not available as a service on Trusty at this time.
Configure riak.conf to match testing configuration.
Remove HHVM due to how much slower the VM takes to spin up.
Add PHP 7.1 but allow it to fail since it is still in beta.
Retain config for 5.4 due to high installation market share at the moment.